feat: policy_fast.py, model_fast.py, and their tests
