+++ tests/test_enforcer.py
@@ -435,64 +435,3 @@ class TestConfigSynced(TestConfig):
         # thread needs a moment to exit
         time.sleep(10 / 1000)
         self.assertFalse(e.is_auto_loading_running())
-
-
-class TestFastEnforcer(TestCaseBase):
-    def test_creates_proper_policy(self) -> None:
-        e = self.get_enforcer(
-            get_examples("basic_model.conf"),
-            get_examples("basic_policy.csv"),
-            [2, 1],
-        )
-
-        assert isinstance(e.model.model["p"]["p"].policy, casbin.model.FilterablePolicy)
-
-    def test_initializes_model(self) -> None:
-        e = self.get_enforcer(
-            get_examples("basic_model.conf"),
-            get_examples("basic_policy.csv"),
-            [2, 1],
-        )
-
-        assert list(e.model.model["p"]["p"].policy) == [
-            ["alice", "data1", "read"],
-            ["bob", "data2", "write"],
-        ]
-
-    def test_able_to_clear_policy(self) -> None:
-        e = self.get_enforcer(
-            get_examples("basic_model.conf"),
-            get_examples("basic_policy.csv"),
-            [2, 1],
-        )
-
-        e.clear_policy()
-
-        assert isinstance(e.model.model["p"]["p"].policy, casbin.FilterablePolicy)
-        assert list(e.model.model["p"]["p"].policy) == []
-
-    def test_able_to_enforce_rule(self) -> None:
-        e = self.get_enforcer(
-            get_examples("basic_model.conf"),
-            get_examples("basic_policy.csv"),
-            [2, 1],
-        )
-
-        assert e.enforce("alice", "data1", "read")
-        assert not e.enforce("alice2", "data1", "read")
-
-    def test_speed_of_fast_enforcer(self) -> None:
-        e = self.get_enforcer(
-            get_examples("basic_model.conf"),
-            get_examples("basic_policy.csv"),
-            [2, 1],
-        )
-
-        for x in range(100):
-            for y in range(1000):
-                e.add_policy(f"/user{x}", f"/obj{y}", "read")
-
-        s = time.time()
-        # this is the absolute worst case last entry and should require iterating 10M rows and be very slow
-        a = e.enforce("/user99", "/obj999", "read")
-        print(a, (time.time() - s) * 1000)
